Struct gstreamer_rtsp::RTSPUrl

pub struct RTSPUrl(_);

Provides helper functions to handle RTSP urls.

Methods

impl RTSPUrl

pub fn decode_path_components(&self) -> Vec<String>

Splits the path of self on '/' boundaries, decoding the resulting components,

The decoding performed by this routine is "URI decoding", as defined in RFC 3986, commonly known as percent-decoding. For example, a string "foo\%2fbar" will decode to "foo/bar" -- the \%2f being replaced by the corresponding byte with hex value 0x2f. Note that there is no guarantee that the resulting byte sequence is valid in any given encoding. As a special case, \%00 is not unescaped to NUL, as that would prematurely terminate the string.

Also note that since paths usually start with a slash, the first component will usually be the empty string.

Returns

None-terminated array of URL components. Free with g_strfreev when no longer needed.

pub fn get_request_uri(&self) -> Option<String>

Get a newly allocated string describing the request URI for self.

Returns

a string with the request URI. g_free after usage.

pub fn set_port(&mut self, port: u16) -> RTSPResult

Set the port number in self to port.

port

the port

Returns

RTSPResult::Ok.

pub fn parse(urlstr: &str) -> (RTSPResult, RTSPUrl)

Parse the RTSP urlstr into a newly allocated RTSPUrl. Free after usage with RTSPUrl::free.

urlstr

the url string to parse

url

location to hold the result.

Returns

a RTSPResult.